refactor: :recycle: update component names
