--- src/components/CustomHeader/CustomHeader.story.tsx
@@ -1,8 +1,0 @@
-import CustomHeader from './CustomHeader'
-import { navigation } from '../../routes/navigation'
-
-export default {
-  title: 'Header',
-}
-
-export const Usage = () => <CustomHeader links={navigation} />
