+++ src/components/Header/Header.story.tsx
@@ -1,0 +1,8 @@
+import Header from './Header'
+import { navigation } from '../../routes/navigation'
+
+export default {
+  title: 'Header',
+}
+
+export const Usage = () => <Header links={navigation} />
