+++ src/components/Header/Header.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen } from '@testing-library/react'
-import CustomHeader from './CustomHeader'
+import Header from './Header'
 import { navigation } from '../../routes/navigation'
 
 describe('Welcome component', () => {
   it('has correct Next.js theming section link', () => {
-    render(<CustomHeader links={navigation} />)
+    render(<Header links={navigation} />)
     expect(screen.getByText('this guide')).toHaveAttribute(
       'href',
       'https://mantine.dev/theming/next/'
